SDL2 -> SDL3
SDL3 is built from source by default and optionally statically linked; both of these are controlled using Cargo features in the CLI and GUI packages

AppImage builds need to statically link SDL3 when building SDL3 from source because otherwise the AppImage build process won't locate the SDL3 shared library file (possibly a way to fix this but I didn't see anything obvious)

these changes required removing two config settings:
* borderless vs. exclusive fullscreen, since changing this seems to be much more difficult in SDL3
* audio hardware/device queue size, which SDL3 no longer seems to expose directly

+++ common/jgenesis-common/src/input.rs
@@ -7,15 +7,15 @@ pub enum Player {
 #[inline]
 #[must_use]
 pub fn viewport_position_to_frame_position(
-    x: i32,
-    y: i32,
+    x: f32,
+    y: f32,
     frame_size: FrameSize,
     display_area: DisplayArea,
 ) -> Option<(u16, u16)> {
-    let display_left = display_area.x as i32;
-    let display_right = display_left + display_area.width as i32;
-    let display_top = display_area.y as i32;
-    let display_bottom = display_top + display_area.height as i32;
+    let display_left = display_area.x as f32;
+    let display_right = display_left + display_area.width as f32;
+    let display_top = display_area.y as f32;
+    let display_bottom = display_top + display_area.height as f32;
 
     if !(display_left..display_right).contains(&x) || !(display_top..display_bottom).contains(&y) {
         return None;
